test(maestro): trace ART signature frames
